32 bit driver support     For change in CosGpuCommand.h : KMD need to support both 32 bit and 64 bit UMDs

+++ compute-only-sample/coscommon/CosGpuCommand.h
@@ -43,7 +43,7 @@ struct GpuResourceCopy
 {
     PHYSICAL_ADDRESS    m_dstGpuAddress;
     PHYSICAL_ADDRESS    m_srcGpuAddress;
-    size_t              m_sizeBytes;
+    UINT                m_sizeBytes;
 };
 
 //
